portal/requirements.txt
chainlit==2.0.4
psycopg[binary]==3.1.18
httpx==0.27.0
pyyaml==6.0.1
boto3==1.34.0
google-generativeai==0.8.3
